Move core camera preview stuff from CameraKitView into isolated CameraPreview
